Don't interpolate bad things from the parent profiles
Parent profiles might declare things like modules that should not be
merged in.

It's actually unclear if I should be merging in anything from parents,
since they should get inherited.  Rather, it's possible that I should
only be pulling property values out of the parents.  But this is a
semi-minimal change for now.